Use HMACCosClient to presign the ignition URL
